(svn r4698) - When selecting files to commit, select the correct ones... missed header file update in r4697

+++ newgrf_engine.h
@@ -25,7 +25,7 @@ SpriteID GetRotorOverrideSprite(EngineID engine, const Vehicle* v);
 void SetEngineGRF(EngineID engine, uint32 grfid);
 uint32 GetEngineGRFID(EngineID engine);
 
-uint16 GetVehicleCallback(byte callback, uint32 param1, uint32 param2, EngineID engine, const Vehicle *v);
+uint16 GetVehicleCallback(uint16 callback, uint32 param1, uint32 param2, EngineID engine, const Vehicle *v);
 bool UsesWagonOverride(const Vehicle *v);
 #define GetCustomVehicleSprite(v, direction) GetCustomEngineSprite(v->engine_type, v, direction)
 #define GetCustomVehicleIcon(et, direction) GetCustomEngineSprite(et, NULL, direction)
